Refactored stream creation.
